+++ src/ui/articles_list/view.rs
@@ -266,9 +266,14 @@ impl<'a> ArticleListViewData<'a> {
                 let row_vec: Vec<Line> = placeholders
                     .iter()
                     .map(|placeholder| match *placeholder {
-                        "{title}" => {
-                            html_sanitize(article.title.as_deref().unwrap_or("no title")).into()
-                        }
+                        "{title}" => html_sanitize(
+                            article
+                                .title
+                                .as_deref()
+                                .or(article.summary.as_deref())
+                                .unwrap_or("no title and summary"),
+                        )
+                        .into(),
                         "{tag_icons}" => Line::from(
                             match model_data.tags_for_article().get(&article.article_id) {
                                 Some(tag_ids) => {